refactor: 원화를 달러화로 변경

+++ src/main/java/com/DahnTa/dto/response/StockGameResultResponse.java
@@ -3,10 +3,10 @@ package com.DahnTa.dto.response;
 import lombok.Builder;
 
 @Builder
-public record StockGameResultResponse(double finalReturnRate, int initialFunds, int finalAmount) {
+public record StockGameResultResponse(double finalReturnRate, double initialFunds, double finalAmount) {
 
-    public static StockGameResultResponse create(double finalReturnRate, int initialFunds,
-        int finalAmount) {
+    public static StockGameResultResponse create(double finalReturnRate, double initialFunds,
+        double finalAmount) {
 
         return StockGameResultResponse.builder()
             .finalReturnRate(finalReturnRate)
