+++ src/main/java/com/DahnTa/entity/Possession.java
@@ -6,6 +6,7 @@ import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
 import jakarta.persistence.OneToOne;
 import jakarta.persistence.Table;
 import lombok.Builder;
@@ -24,11 +25,9 @@ public class Possession {
     @JoinColumn(name = "stock_id", nullable = false)
     private Stock stock;
 
-    /*
     @ManyToOne
     @JoinColumn(name = "user_id", nullable = false)
     private User user;
-     */
 
     @Column(name = "quantity", nullable = false)
     private int quantity;
@@ -59,7 +58,7 @@ public class Possession {
         this.quantity -= quantity;
     }
 
-    public int calculateAmount(int stockCurrentPrice) {
+    public double calculateAmount(double stockCurrentPrice) {
         return this.getQuantity() * stockCurrentPrice;
     }
